backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
google-genai
pypdf
python-multipart
requests
trafilatura
beautifulsoup4
lxml
